
\name{areTheSame}

\alias{areTheSame}

\title{Are they the same mathematical entity?}

\description{Check whether two objects are the same mathematical entity.}

\usage{areTheSame(x, y)}

\arguments{
  \item{x}{an R object representing a mathematical entity}
  \item{y}{another R object possibly representing the same mathematical entity}
}

\details{The function \code{areTheSame} is generic.}

\value{Returns a logical value telling whether the two objects are the same mathematical entity or not.}

% \references{ ... }

\author{Jens Henrik Badsberg, Claus Dethlefsen, Luca La Rocca}

% \note{ ~~further notes~~ }

% \section{Warning }{....} ~

%\seealso{...}

% \examples{}

\keyword{methods}
