
\name{card}

\alias{card}
\alias{card,vector-method}

\title{Get the cardinality of an object}

\description{Gets the cardinality of an object.}

\usage{card(object, ...)}

\arguments{
  \item{object}{an R object whose cardinality is to be retrieved}
  \item{\dots}{additional parameters to be used when retrieving cardinality}
}

\details{The function \code{card} is generic.}

\value{Returns a numeric value corresponding to the cardinality of \code{object}.}

% \references{ ... }

\author{Jens Henrik Badsberg, Claus Dethlefsen, Luca La Rocca}

\note{For vectors cardinality is the same as length.}

% \section{Warning }{....} ~

%\seealso{...}

% \examples{}

\keyword{methods}
