
\name{giRaph}

\alias{giRaph}

\title{The package `giRaph': summary information}

\description{This package provides classes and methods for graph representation and manipulation in R}

\section{Details}{
\itemize{
  \item 'giRaph' provides a general framework for dealing with mathematical graphs.
  The setting is very general and includes four types of representations
  and four types of graphs with conversions between them.
  \item Other packages for mathematical graphs in R include: Bioconductor 'graph', 'mathgraph', 'dynamicGraph', 'ggm'.
  }

The package is inteded as a contribution to the gR-project described by Lauritzen (2002).
}

\section{Authors}{
Jens Henrik Badsberg, Biometry Research Unit, Danish Institute of
Agricultural Sciences, DK-8830 Tjele, Denmark

Claus Dethlefsen, Center for Cardiovascular Research, Aalborg Hospital,
Aarhus University Hospital, DK-9000 Aalborg, Denmark

Luca La Rocca, Department of Social, Cognitive and Quantitative Sciences,
University of Modena and Reggio Emilia, IT-42100 Reggio Emilia, Italy.
}

\references{
Lauritzen, S. L. (2002). gRaphical Models in R. \emph{R News}, 3(2)39.
}

\keyword{models}
\keyword{graphs}
