
\name{isEmpty}

\alias{isEmpty}

\alias{isEmpty,vector-method}
\alias{isEmpty,NULL-method}

\title{Is the object empty?}

\description{Check whether an object is empty.}

\usage{isEmpty(object, ...)}

\arguments{
  \item{object}{an object to be checked}
  \item{\dots}{additional parameters to be used when checking the object}
}

\details{The function \code{isEmpty} is generic.}

\value{Returns a logical value telling whether the object is empty or not.}

% \references{ ... }

\author{Jens Henrik Badsberg, Claus Dethlefsen, Luca La Rocca}

% \note{ ~~further notes~~ }

% \section{Warning }{....} ~

%\seealso{...}

% \examples{}

\keyword{methods}
