
\name{isPresent}

\alias{isPresent}

\title{Is the first object present in the second one?}
\description{Check whether the first object is present in the second one.}

\usage{isPresent(el, ou)}

\arguments{
  \item{el}{element to look for}
  \item{ou}{place where to look for it}
}

\details{The function \code{isPresent} is generic.}

\value{Returns a logical value telling whether the first object is present in the second one or not.}

%\references{ ~put references to the literature/web site here ~ }

\author{Jens Henrik Badsberg, Claus Dethlefsen, Luca La Rocca}

% \note{ ~~further notes~~ }

% \section{Warning }{....} ~

%\seealso{...}

% \examples{}

\keyword{methods}
