
\name{maxId}

\alias{maxId}

\title{Get the maximum numeric identifier of an object}

\description{Gets the maximum numeric identifier of an object.}

\usage{maxId(x)}

\arguments{
  \item{x}{an R object with numeric identifiers}
}

\details{The function \code{maxId} is generic.}

\value{Returns a numeric value corresponding to the maximum numeric identifier of \code{x}.}

% \references{ ... }

\author{Jens Henrik Badsberg, Claus Dethlefsen, Luca La Rocca}

% \note{ ~~further notes~~ }

% \section{Warning }{....} ~

%\seealso{...}

% \examples{}

\keyword{methods}
