
\name{showRel}

\alias{showRel}

\title{Function to show an object relative to a given code}

\description{Shows an object relative to a given code.}

\usage{showRel(object,code)}

\arguments{
  \item{object}{an R object to be shown}
  \item{code}{object containing the code to be used}
}

\details{The function \code{showRel} is generic.}

\value{Returns an invisible \code{NULL}.}

% \references{ ... }

\author{Jens Henrik Badsberg, Claus Dethlefsen, Luca La Rocca}

% \section{Warning }{....} ~

%\seealso{...}

% \examples{}

\keyword{methods}
