\name{giRaph}
\alias{giRaph}
\title{
The package `giRaph': summary information
}
\description{
  This package provides graph representation and manipulation in R
}
\section{Details}{
\itemize{
  \item giRaph provides the general framework for setting up
  representations for mathematical graphs. The setting is general and
  include four types of representations and four types of graphs with
  conversions between them.
  \item Other packages for mathematical graphs in R includes:
  Bioconductors 'graph', mathgraph, dynamicGraph, ggm.
  }
The package is a contribution to the gR-project 
described by Lauritzen (2002).
}
\section{Authors}{
Jens Henrik Badsberg, Biometry Research Unit, Danish Institute of
Agricultural Sciences, DK-8830 Tjele, Denmark

Claus Dethlefsen, Center for Cardiovascular Research, Aalborg Hospital,
Aarhus University Hospital, DK-9000 Aalborg, Denmark

Luca La Rocca, Centro di Ricerca sulla Societa della Comunicazione e
dell'Informazione, Universita degli studi di Modena e Reggio Emilia,
42100 Reggio Emilia, Italia. 
}
\references{
Lauritzen, S. L. (2002). gRaphical Models in R.
\emph{R News}, 3(2)39.
}
\keyword{models}
\keyword{graphs}
