
\name{recode}

\alias{recode}

\title{Function to recode an object from a given source code to a given target code}

\description{Recodes an object from a given source code to a given target code}

\usage{recode(object, src, dst)}

\arguments{
  \item{object}{ an object to be recoded }
  \item{src}{ an object containing the source code }
  \item{dst}{ an object containing the target code }
}

\details{The function \code{recode} is generic.}

\value{The recoded object is returned.}

%\references{ ~put references to the literature/web site here ~ }

\author{Jens Henrik Badsberg, Claus Dethlefsen, Luca La Rocca}

% \note{ ~~further notes~~ }

% \section{Warning }{....} ~

%\seealso{...}

% \examples{}

\keyword{methods}
